Add checkForComplete() to Game.java, refactor Square.java, and add allClear to Board.java
Square has two static variables spacesFlagged and incorrectFlags which count the number of spaces flagged and keep track if all flagged spaces are bombs.

allClear() method in board makes sure all spaces are either flagged or showing.

checkForComplete() uses these methods to check if a win condition is met whenever a space if flagged or revealed.